[MergeDups] Reorganize
