+++ src/components/GoalTimeline/tests/GoalReducer.test.tsx
@@ -6,7 +6,7 @@ import {
 } from "components/GoalTimeline/Redux/GoalReduxTypes";
 import { CreateCharInv } from "goals/CharacterInventory/CharacterInventoryTypes";
 import { HandleFlags } from "goals/HandleFlags/HandleFlags";
-import { MergeDups } from "goals/MergeDupGoal/MergeDups";
+import { MergeDups } from "goals/MergeDuplicates/MergeDupsTypes";
 import { ReviewEntries } from "goals/ReviewEntries/ReviewEntries";
 import { SpellCheckGloss } from "goals/SpellCheckGloss/SpellCheckGloss";
 import { ValidateChars } from "goals/ValidateChars/ValidateChars";
