+++ src/goals/MergeDuplicates/MergeDupsCompleted.tsx
@@ -8,8 +8,8 @@ import { Flag, MergeUndoIds, Sense, Word } from "api/models";
 import { getFrontierWords, getWord, undoMerge } from "backend";
 import { FlagButton } from "components/Buttons";
 import { CancelConfirmDialog } from "components/Dialogs";
-import SenseCardContent from "goals/MergeDupGoal/MergeDupStep/SenseCardContent";
-import { MergesCompleted } from "goals/MergeDupGoal/MergeDupsTypes";
+import SenseCardContent from "goals/MergeDuplicates/MergeDupsStep/SenseCardContent";
+import { MergesCompleted } from "goals/MergeDuplicates/MergeDupsTypes";
 import { StoreState } from "types";
 import theme from "types/theme";
 import { newFlag } from "types/word";
