+++ src/goals/MergeDuplicates/MergeDupsStep/SenseCardContent.tsx
