+++ src/goals/MergeDuplicates/MergeDupsTypes.ts
@@ -1,4 +1,5 @@
 import { MergeSourceWord, MergeUndoIds, MergeWords, Word } from "api/models";
+import { Goal, GoalName, GoalType } from "types/goals";
 
 export interface MergeDupData {
   plannedWords: Word[][];
@@ -18,4 +19,13 @@ export function newMergeWords(
   deleteOnly = false
 ): MergeWords {
   return { parent, children, deleteOnly };
+}
+
+export class MergeDups extends Goal {
+  constructor(
+    steps: MergeStepData[] = [],
+    data: MergeDupData = { plannedWords: [[]] }
+  ) {
+    super(GoalType.MergeDups, GoalName.MergeDups, steps, data);
+  }
 }
