+++ src/rootReducer.ts
@@ -8,7 +8,7 @@ import { createProjectReducer } from "components/ProjectScreen/CreateProject/Red
 import { pronunciationsReducer } from "components/Pronunciations/Redux/PronunciationsReducer";
 import { treeViewReducer } from "components/TreeView/Redux/TreeViewReducer";
 import { characterInventoryReducer } from "goals/CharacterInventory/Redux/CharacterInventoryReducer";
-import { mergeDupStepReducer } from "goals/MergeDupGoal/Redux/MergeDupReducer";
+import { mergeDupStepReducer } from "goals/MergeDuplicates/Redux/MergeDupsReducer";
 import { reviewEntriesReducer } from "goals/ReviewEntries/ReviewEntriesComponent/Redux/ReviewEntriesReducer";
 import { StoreState } from "types";
 import { analyticsReducer } from "types/Redux/analytics";
